Changed paths: M       INSTALL M       Include/CAT_Curvature.h M       Include/CAT_Surf.h A       Include/CAT_SurfParameters.h M       Lib/CAT_Curvature.c M       Lib/CAT_Gyrification.c M       Lib/CAT_Surf.c A       Lib/CAT_SurfParameters.c M       Lib/CAT_Topology.c M       Makefile.am M       Progs/CAT_DumpConvexity.c M       Progs/CAT_Hausdorff.c M       Progs/CAT_Sharpness.c M       Progs/CAT_SulcusDepth.c D     C Progs/CAT_Surf2Surf.c       >   local unversioned, incoming add upon update M       depcomp M       gifticlib/gifti_io.c M       mkinstalldirs
(1) Created CAT_SurfParameters and moved some functions from CAT_Surf.c
    and CAT_Curvature.c to this function that covers estimation of
    different surface parameters.
(2) Renamed some functions from get_ and calc_ to compute_
(3) Resolved some compiler warnings.



git-svn-id: svn://141.35.69.214/CAT@343 85a05d3d-4204-4b9d-9be8-300d97d0f21f

+++ gifticlib/gifti_io.c
@@ -3354,7 +3354,7 @@ long long gifti_compare_raw_data(const void * p1, const void * p2,
 
 /* make a local definition for this symmetric fractional difference */
 #undef GIFTI_SFD
-#define GIFTI_SFD(a,b) (fabs((a)-(double)(b))/(fabs(a)+fabs(b)))
+#define GIFTI_SFD(a,b) (fabs((double)a-(double)b)/(fabs((double)a)+fabs((double)b)))
 
 /*---------------------------------------------------------------------*/
 /*! approximate comparison of raw data, returing the first location difference
